@@ -1,14 +1,18 @@
 import { CssReset } from '@dhis2/ui'
 import React from 'react'
-import ReactDOM from 'react-dom'
+import { createRoot } from 'react-dom/client'
 import App from './App.jsx'
 import 'typeface-roboto'
 import './index.css'
 
-ReactDOM.render(
+const container = document.getElementById('dhis2-app-root')
+const root = createRoot(container)
+root.render(
     <>
         <CssReset />
         <App />
-    </>,
-    document.getElementById('dhis2-app-root')
+    </>
 )
+
+// ToDo: remove before merging
+console.log('     ::: using react@18 :::')
